GA/NT: warnings
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@63105 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Additions/WINNT/Graphics/Video/disp/wddm/VBoxDispD3D.cpp
@@ -644,7 +644,7 @@ static void vboxWddmDbgSynchMemCheck(PVBOXWDDMDISP_ALLOCATION pAlloc, D3DLOCKED_
     Assert(pAlloc->pvMem);
     int iRc = 0;
 
-    if (pAlloc->SurfDesc.pitch == pLockInfo->Pitch)
+    if (pAlloc->SurfDesc.pitch == (UINT)pLockInfo->Pitch)
     {
         Assert(pAlloc->SurfDesc.cbSize);
         iRc = memcmp(pLockInfo->pBits, pAlloc->pvMem, pAlloc->SurfDesc.cbSize);
